@@ -18,18 +18,18 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.2
-  flutter_secure_storage: ^3.3.5
-  flutter_bloc: ^6.1.3
-  equatable: ^1.2.6
-  meta: ^1.2.4
-  flutter_slidable: ^0.5.7
-  bottom_navy_bar: ^5.6.0
+  http: ^0.13.0
+  flutter_secure_storage: ^4.1.0
+  flutter_bloc: ^7.0.0 
+  equatable: ^2.0.0
+  meta: ^1.3.0
+  flutter_slidable: ^0.6.0-nullsafety.0
+
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
