@@ -1,3 +1,3 @@
 //final String SERVER_URL = 'http://8ge2e.mocklab.io';
 //final String SERVER_URL = 'http://192.168.1.174:8080';
-final String SERVER_URL = 'http://svf.ddns.net:8088';
+const ServerAddress = 'svf.ddns.net:8088';
